Move legacy intrinsic gas to basic check
